refactor(app): Update app.dart to use multi-repository and multi-bloc providers
This commit updates the app.dart file to use the MultiRepositoryProvider and MultiBlocProvider widgets for managing dependencies. It also adds the StoreService and CacheService as dependencies to the Shop4MeApp widget. The NavigationBloc and ProductSearchBloc are now created using the storeService. Additionally, the Hive initialization and cacheService initialization are moved to the main.dart file.

Refactor app.dart to use multi-repository and multi-bloc providers, and update dependencies in main.dart